UNT-T20517: Docs update & Code refactor
