+++ Sources/SSStoryStatus/SSStoryStatus.docc/SSStoryStatus.md
@@ -1,6 +1,6 @@
 # ``SSStoryStatus``
 
-Creates Profile listing and Story view like WhatsApp/Instagram with customizable components. 
+Creates Profile listing and Story view like WhatsApp/Instagram with customizable components.
 
 ## Overview
 
@@ -9,6 +9,7 @@ Creates Profile listing and Story view like WhatsApp/Instagram with customizable
 SSStoryStatus is a versatile and intuitive SwiftUI library designed to effortlessly display user lists and seamlessly showcase their captivating stories. This library empowers developers to effortless integration of user listings with story viewing functionality. This library provides complete control over view components for UI customization.
 
 ## Feature
+
 - Image & Video stories
 - Customizable component with styles
 - Built-in caching support
